+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Changelog
 * Add new method :meth:`pyvcf.VcfFrame.from_string`.
 * Remove ``nrows`` argument from :meth:`pyvcf.VcfFrame.from_file` method.
 * Add new argument 'regions' to :meth:`pyvcf.VcfFrame.from_file` method.
+* Add new property ``pybed.BedFrame.shape``.
+* Add new method :meth:`pybed.BedFrame.regions`.
 
 0.25.0 (2021-10-09)
 -------------------